Adds editing cards in card list
It works in a bit of a wonky way

The first card isn't linked to a card on the list until you hit add card. Then the card that is added can be edited. If you hit add card while a card is being edited (so any time other than the first time) it will create a copy of the current card and begin editing it. To swap to a different card, click on it in the list.